fix: consolidate provider-session tests for pkg-file-count compliance
The mergeability split into provider_session_detail_http_test.go pushed pkg/api over the 15-file limit; move regressions into the existing consolidated server_factory_sessions_test.go instead.
